feat: add Commit hashes to release notes
- adds short version of commit hash to release notes
- moves Commit into commit.go
- adds Hash as a type with String() method
